requirements.txt
Flask==1.1.4
flask-mysqldb==0.2.0
secrets==1.0.2
Flask-Cors==3.0.10
PyJWT==2.1.0
Flask-Bcrypt==0.7.1
